+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@
 - added: `git machete help <command>` now accepts command aliases
 - fixed: removed stack trace when terminating `git machete` command prompt with `Ctrl+D`
 - added: support for Python 3.11
+- added: machete config key `machete.traverse.push` that controls default behavior of `traverse` command
 
 ## New in git-machete 3.12.5